move MRDivEntity
