+++ src/core/entities/MRDivEntity.js
@@ -2,6 +2,9 @@ import * as THREE from 'three';
 
 import { MREntity } from 'mrjs/core/MREntity';
 
+import { MRDivEntity } from 'mrjs/core/MRDivEntity';
+import { MRDivEntity } from 'mrjs/core/entities/MRDivEntity';
+
 import { mrjsUtils } from 'mrjs';
 
 /**
